refactor(package): complete package structure alignment with target spec
Structural changes
- models/segmentation_head.py → models/heads/segmentation.py (canonical);
  old path converted to deprecation shim; all internal imports updated
- Add models/heads/detection.py with DetectionHead class (bbox + class embed)
- Add datasets/coco_eval.py (CocoEvaluator extracted from export/benchmark.py)
- Add datasets/base.py (DatasetBuilderProtocol)
- Add export/protocols.py (ExporterProtocol)
- Add training/trainer.py (build_trainer() extracted from training/__init__.py)
- Rename platform/platform_downloads.py → platform/downloads.py

export/ cleanup
- Rename export/export.py → export/main.py (eliminates rfdetr.export.export oddity)
- Split export/_onnx/optimizer.py + export logic → export/_onnx/exporter.py;
  optimizer.py was a new file (not in v1.5.2), deleted as shim artifact
- Add export/tensorrt.py (trtexec/parse_trtexec_output extracted from export.py)
- Add export/requirements.txt content to pyproject.toml as [deploy] extra;
  delete the file

deploy/ cleanup
- Collapse deploy/ to a single __init__.py shim; delete all submodule files
  (benchmark.py, export.py, _onnx/, requirements.txt)
- Consolidate all DeprecationWarning into deploy/__init__.py; use sys.modules
  injection so rfdetr.deploy.benchmark/export remain importable without files
