+++ src/rfdetr/deploy/__init__.py
@@ -6,6 +6,7 @@
 
 """Backward-compatibility shim — rfdetr.deploy is deprecated; use rfdetr.export."""
 
+import sys
 import warnings
 
 warnings.warn(
@@ -15,3 +16,10 @@ warnings.warn(
 )
 
 from rfdetr.export import *  # noqa: F401, F403, E402
+
+# Make old submodule paths still importable without submodule files
+import rfdetr.export.benchmark as _benchmark  # noqa: E402
+import rfdetr.export.main as _export_main  # noqa: E402
+
+sys.modules.setdefault("rfdetr.deploy.benchmark", _benchmark)
+sys.modules.setdefault("rfdetr.deploy.export", _export_main)
