--- src/rfdetr/export/requirements.txt
@@ -1,8 +1,0 @@
-pycuda
-onnx
-onnxsim
-onnxruntime
-onnxruntime-gpu
-onnx_graphsurgeon
-tensorrt>=8.6.1
-polygraphy
