+++ src/rfdetr/models/heads/__init__.py
@@ -4,4 +4,22 @@
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
 
-"""Detection and segmentation head subpackage (stub -- contents extracted in a future phase)."""
+"""Detection and segmentation head subpackage."""
+
+from rfdetr.models.heads.segmentation import (  # noqa: F401
+    DepthwiseConvBlock,
+    MLPBlock,
+    SegmentationHead,
+    calculate_uncertainty,
+    get_uncertain_point_coords_with_randomness,
+    point_sample,
+)
+
+__all__ = [
+    "SegmentationHead",
+    "DepthwiseConvBlock",
+    "MLPBlock",
+    "point_sample",
+    "get_uncertain_point_coords_with_randomness",
+    "calculate_uncertainty",
+]
