+++ src/rfdetr/models/segmentation_head.py
@@ -4,290 +4,22 @@
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
 
+"""Backward-compatibility shim — rfdetr.models.segmentation_head is deprecated; use rfdetr.models.heads.segmentation."""
 
-from typing import Callable
+import warnings
 
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
+warnings.warn(
+    "rfdetr.models.segmentation_head is deprecated; use rfdetr.models.heads.segmentation instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-
-class DepthwiseConvBlock(nn.Module):
-    r"""Simplified ConvNeXt block without the MLP subnet"""
-
-    def __init__(self, dim, layer_scale_init_value=0):
-        super().__init__()
-        self.dwconv = nn.Conv2d(dim, dim, kernel_size=3, padding=1, groups=dim)  # depthwise conv
-        self.norm = nn.LayerNorm(dim, eps=1e-6)
-        self.pwconv1 = nn.Linear(dim, dim)  # pointwise/1x1 convs, implemented with linear layers
-        self.act = nn.GELU()
-        self.gamma = (
-            nn.Parameter(layer_scale_init_value * torch.ones((dim)), requires_grad=True)
-            if layer_scale_init_value > 0
-            else None
-        )
-
-    def _depthwise_conv(self, x: torch.Tensor) -> torch.Tensor:
-        # Always run this depthwise conv with cuDNN disabled to avoid
-        # backend engine selection failures on some CUDA stacks (e.g. T4/Colab).
-        with torch.backends.cudnn.flags(enabled=False):
-            return self.dwconv(x)
-
-    def forward(self, x):
-        input = x
-        x = self._depthwise_conv(x)
-        x = x.permute(0, 2, 3, 1)  # (N, C, H, W) -> (N, H, W, C)
-        x = self.norm(x)
-        x = self.pwconv1(x)
-        x = self.act(x)
-        if self.gamma is not None:
-            x = self.gamma * x
-        x = x.permute(0, 3, 1, 2)  # (N, H, W, C) -> (N, C, H, W)
-
-        return x + input
-
-
-class MLPBlock(nn.Module):
-    def __init__(self, dim, layer_scale_init_value=0):
-        super().__init__()
-        self.norm_in = nn.LayerNorm(dim)
-        self.layers = nn.ModuleList(
-            [
-                nn.Linear(dim, dim * 4),
-                nn.GELU(),
-                nn.Linear(dim * 4, dim),
-            ]
-        )
-        self.gamma = (
-            nn.Parameter(layer_scale_init_value * torch.ones((dim)), requires_grad=True)
-            if layer_scale_init_value > 0
-            else None
-        )
-
-    def forward(self, x):
-        input = x
-        x = self.norm_in(x)
-        for layer in self.layers:
-            x = layer(x)
-        if self.gamma is not None:
-            x = self.gamma * x
-        return x + input
-
-
-class SegmentationHead(nn.Module):
-    def __init__(self, in_dim, num_blocks: int, bottleneck_ratio: int = 1, downsample_ratio: int = 4):
-        super().__init__()
-
-        self.downsample_ratio = downsample_ratio
-        self.interaction_dim = in_dim // bottleneck_ratio if bottleneck_ratio is not None else in_dim
-        self.blocks = nn.ModuleList([DepthwiseConvBlock(in_dim) for _ in range(num_blocks)])
-        self.spatial_features_proj = (
-            nn.Identity() if bottleneck_ratio is None else nn.Conv2d(in_dim, self.interaction_dim, kernel_size=1)
-        )
-
-        self.query_features_block = MLPBlock(in_dim)
-        self.query_features_proj = (
-            nn.Identity() if bottleneck_ratio is None else nn.Linear(in_dim, self.interaction_dim)
-        )
-
-        self.bias = nn.Parameter(torch.zeros(1), requires_grad=True)
-
-        self._export = False
-
-    def export(self):
-        self._export = True
-        self._forward_origin = self.forward
-        self.forward = self.forward_export
-        for name, m in self.named_modules():
-            if hasattr(m, "export") and isinstance(m.export, Callable) and hasattr(m, "_export") and not m._export:
-                m.export()
-
-    def forward(
-        self,
-        spatial_features: torch.Tensor,
-        query_features: list[torch.Tensor],
-        image_size: tuple[int, int],
-        skip_blocks: bool = False,
-    ) -> list[torch.Tensor]:
-        # spatial features: (B, C, H, W)
-        # query features: [(B, N, C)] for each decoder layer
-        # output: (B, N, H*r, W*r)
-        target_size = (image_size[0] // self.downsample_ratio, image_size[1] // self.downsample_ratio)
-        spatial_features = F.interpolate(spatial_features, size=target_size, mode="bilinear", align_corners=False)
-
-        mask_logits = []
-        if not skip_blocks:
-            for block, qf in zip(self.blocks, query_features):
-                spatial_features = block(spatial_features)
-                spatial_features_proj = self.spatial_features_proj(spatial_features)
-                qf = self.query_features_proj(self.query_features_block(qf))
-                mask_logits.append(torch.einsum("bchw,bnc->bnhw", spatial_features_proj, qf) + self.bias)
-        else:
-            assert len(query_features) == 1, "skip_blocks is only supported for length 1 query features"
-            qf = self.query_features_proj(self.query_features_block(query_features[0]))
-            mask_logits.append(torch.einsum("bchw,bnc->bnhw", spatial_features, qf) + self.bias)
-
-        return mask_logits
-
-    def sparse_forward(
-        self,
-        spatial_features: torch.Tensor,
-        query_features: list[torch.Tensor],
-        image_size: tuple[int, int],
-        skip_blocks: bool = False,
-    ) -> list[torch.Tensor]:
-        # spatial features: (B, C, H, W)
-        # query features: [(B, N, C)] for each decoder layer
-        # output: dict containing the intermediate results
-        target_size = (image_size[0] // self.downsample_ratio, image_size[1] // self.downsample_ratio)
-        spatial_features = F.interpolate(spatial_features, size=target_size, mode="bilinear", align_corners=False)
-
-        # num_points = max(spatial_features.shape[-2], spatial_features.shape[-2] * spatial_features.shape[-1] // 16)
-
-        output_dicts = []
-
-        if not skip_blocks:
-            for block, qf in zip(self.blocks, query_features):
-                spatial_features = block(spatial_features)
-                spatial_features_proj = self.spatial_features_proj(spatial_features)
-                qf = self.query_features_proj(self.query_features_block(qf))
-
-                output_dicts.append(
-                    {
-                        "spatial_features": spatial_features_proj,
-                        "query_features": qf,
-                        "bias": self.bias,
-                    }
-                )
-        else:
-            assert len(query_features) == 1, "skip_blocks is only supported for length 1 query features"
-
-            qf = self.query_features_proj(self.query_features_block(query_features[0]))
-
-            output_dicts.append(
-                {
-                    "spatial_features": spatial_features,
-                    "query_features": qf,
-                    "bias": self.bias,
-                }
-            )
-
-        return output_dicts
-
-    def forward_export(
-        self,
-        spatial_features: torch.Tensor,
-        query_features: list[torch.Tensor],
-        image_size: tuple[int, int],
-        skip_blocks: bool = False,
-    ) -> list[torch.Tensor]:
-        assert len(query_features) == 1, "at export time, segmentation head expects exactly one query feature"
-
-        target_size = (image_size[0] // self.downsample_ratio, image_size[1] // self.downsample_ratio)
-        spatial_features = F.interpolate(spatial_features, size=target_size, mode="bilinear", align_corners=False)
-
-        if not skip_blocks:
-            for block in self.blocks:
-                spatial_features = block(spatial_features)
-
-        spatial_features_proj = self.spatial_features_proj(spatial_features)
-
-        qf = self.query_features_proj(self.query_features_block(query_features[0]))
-        return [torch.einsum("bchw,bnc->bnhw", spatial_features_proj, qf) + self.bias]
-
-
-def point_sample(input, point_coords, **kwargs):
-    """
-    A wrapper around :function:`torch.nn.functional.grid_sample` to support 3D point_coords tensors.
-    Unlike :function:`torch.nn.functional.grid_sample` it assumes `point_coords` to lie inside
-    [0, 1] x [0, 1] square.
-
-    Args:
-        input (Tensor): A tensor of shape (N, C, H, W) that contains features map on a H x W grid.
-        point_coords (Tensor): A tensor of shape (N, P, 2) or (N, Hgrid, Wgrid, 2) that contains
-        [0, 1] x [0, 1] normalized point coordinates.
-
-    Returns:
-        output (Tensor): A tensor of shape (N, C, P) or (N, C, Hgrid, Wgrid) that contains
-            features for points in `point_coords`. The features are obtained via bilinear
-            interplation from `input` the same way as :function:`torch.nn.functional.grid_sample`.
-    """
-    add_dim = False
-    if point_coords.dim() == 3:
-        add_dim = True
-        point_coords = point_coords.unsqueeze(2)
-    output = F.grid_sample(input, 2.0 * point_coords - 1.0, padding_mode="border", **kwargs)
-    if add_dim:
-        output = output.squeeze(3)
-    return output
-
-
-def get_uncertain_point_coords_with_randomness(
-    coarse_logits, uncertainty_func, num_points, oversample_ratio=3, importance_sample_ratio=0.75
-):
-    """
-    Sample points in [0, 1] x [0, 1] coordinate space based on their uncertainty. The unceratinties
-        are calculated for each point using 'uncertainty_func' function that takes point's logit
-        prediction as input.
-    See PointRend paper for details.
-
-    Args:
-        coarse_logits (Tensor): A tensor of shape (N, C, Hmask, Wmask) or (N, 1, Hmask, Wmask) for
-            class-specific or class-agnostic prediction.
-        uncertainty_func: A function that takes a Tensor of shape (N, C, P) or (N, 1, P) that
-            contains logit predictions for P points and returns their uncertainties as a Tensor of
-            shape (N, 1, P).
-        num_points (int): The number of points P to sample.
-        oversample_ratio (int): Oversampling parameter.
-        importance_sample_ratio (float): Ratio of points that are sampled via importnace sampling.
-
-    Returns:
-        point_coords (Tensor): A tensor of shape (N, P, 2) that contains the coordinates of P
-            sampled points.
-    """
-    assert oversample_ratio >= 1
-    assert importance_sample_ratio <= 1 and importance_sample_ratio >= 0
-    num_boxes = coarse_logits.shape[0]
-    num_sampled = int(num_points * oversample_ratio)
-    point_coords = torch.rand(num_boxes, num_sampled, 2, device=coarse_logits.device)
-    point_logits = point_sample(coarse_logits, point_coords, align_corners=False)
-    # It is crucial to calculate uncertainty based on the sampled prediction value for the points.
-    # Calculating uncertainties of the coarse predictions first and sampling them for points leads
-    # to incorrect results.
-    # To illustrate this: assume uncertainty_func(logits)=-abs(logits), a sampled point between
-    # two coarse predictions with -1 and 1 logits has 0 logits, and therefore 0 uncertainty value.
-    # However, if we calculate uncertainties for the coarse predictions first,
-    # both will have -1 uncertainty, and the sampled point will get -1 uncertainty.
-    point_uncertainties = uncertainty_func(point_logits)
-    num_uncertain_points = int(importance_sample_ratio * num_points)
-    num_random_points = num_points - num_uncertain_points
-    idx = torch.topk(point_uncertainties[:, 0, :], k=num_uncertain_points, dim=1)[1]
-    shift = num_sampled * torch.arange(num_boxes, dtype=torch.long, device=coarse_logits.device)
-    idx += shift[:, None]
-    point_coords = point_coords.view(-1, 2)[idx.view(-1), :].view(num_boxes, num_uncertain_points, 2)
-    if num_random_points > 0:
-        point_coords = torch.cat(
-            [
-                point_coords,
-                torch.rand(num_boxes, num_random_points, 2, device=coarse_logits.device),
-            ],
-            dim=1,
-        )
-    return point_coords
-
-
-def calculate_uncertainty(logits: torch.Tensor) -> torch.Tensor:
-    """
-    We estimate uncertainty as L1 distance between 0.0 and the logit prediction in 'logits' for the
-        foreground class in `classes`.
-    Args:
-        logits (Tensor): A tensor of shape (R, 1, ...) for class-specific or
-            class-agnostic, where R is the total number of predicted masks in all images and C is
-            the number of foreground classes. The values are logits.
-    Returns:
-        scores (Tensor): A tensor of shape (R, 1, ...) that contains uncertainty scores with
-            the most uncertain locations having the highest uncertainty score.
-    """
-    assert logits.shape[1] == 1
-    gt_class_logits = logits.clone()
-    return -(torch.abs(gt_class_logits))
+from rfdetr.models.heads.segmentation import *  # noqa: F401, F403, E402
+from rfdetr.models.heads.segmentation import (  # noqa: E402
+    DepthwiseConvBlock,
+    MLPBlock,
+    SegmentationHead,
+    calculate_uncertainty,
+    get_uncertain_point_coords_with_randomness,
+    point_sample,
+)
